feat: adding outfits
